feat: シミュレーター排他制御の表示といくつかのリファクタ
