+++ frontend/app/components/modules/simulator-parts/simulator-chair-status-label.tsx
